requirements.txt
alabaster==0.7.12
appdirs==1.4.3
asgiref==3.2.7
aspy.yaml==1.3.0
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
Babel==2.8.0
backcall==0.1.0
bcbio-gff==0.6.6
beautifulsoup4==4.9.0
biopython==1.76
black==19.10b0
bleach==3.1.1
cached-property==1.5.1
certifi==2020.4.5.1
cfgv==3.1.0
chardet==3.0.4
click==7.1.1
colorama==0.4.3
commonmark==0.9.1
ConfigArgParse==1.2.3
coverage==5.1
cssselect==1.1.0
cycler==0.10.0
datrie==0.8.2
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
Django==3.0.4
django-cors-headers==3.2.1
django-environ==0.4.5
django-extensions==2.2.8
django-test-plus==1.1.1
django-url-checks==0.2.0
djangorestframework==3.11.0
djangorestframework-xml==1.4.0
docutils==0.16
entrypoints==0.3
factory-boy==2.11.1
Faker==0.9.2
feedfinder2==0.0.4
feedparser==5.2.1
filelock==3.0.12
flake8==3.7.9
flake8-bugbear==18.2.0
flake8-docstrings==1.3.0
flake8-polyfill==1.0.2
freezegun==0.3.15
gitdb==4.0.4
GitPython==3.1.1
identify==1.4.11
idna==2.9
imagesize==1.2.0
ipdb==0.13.2
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.16.0
jieba3k==0.35.1
Jinja2==2.11.1
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.0.0
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lxml==4.5.0
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.2.0
nbconvert==5.6.1
nbformat==5.0.4
newspaper3k==0.2.8
-e git+https://github.com/anouarh/augur.git@8568194acc5948d3510a0747832239109bf63c30#egg=nextstrain_augur
nltk==3.5
nodeenv==1.3.5
notebook==6.0.3
numpy==1.18.3
packaging==20.3
pandas==0.25.3
pandocfilters==1.4.2
parso==0.6.2
pathspec==0.7.0
pep8-naming==0.7.0
phylo-treetime==0.7.5
pickleshare==0.7.5
Pillow==7.1.2
pluggy==0.13.1
pre-commit==1.10.4
prometheus-client==0.7.1
prompt-toolkit==3.0.4
psutil==5.7.0
py==1.8.1
pycodestyle==2.5.0
pydocstyle==5.0.2
pyflakes==2.1.1
Pygments==2.6.1
pylint==2.4.4
pyparsing==2.4.7
pyrsistent==0.15.7
pytest==5.4.1
pytest-cov==2.8.1
pytest-mock==2.0.0
python-dateutil==2.8.1
pytz==2019.3
pywin32==227
pywinpty==0.5.7
PyYAML==5.3
pyzmq==19.0.0
qtconsole==4.7.1
QtPy==1.9.0
ratelimiter==1.2.0.post0
recommonmark==0.6.0
regex==2020.2.20
requests==2.23.0
requests-file==1.5.1
scipy==1.4.1
Send2Trash==1.5.0
six==1.14.0
smmap==3.0.2
snakemake==5.10.0
snowballstemmer==2.0.0
soupsieve==2.0
Sphinx==2.4.4
sphinx-argparse==0.2.5
sphinx-markdown-tables==0.0.12
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.2
tinysegmenter==0.3
tldextract==2.2.2
toml==0.10.0
toposort==1.5
tornado==6.0.4
tqdm==4.46.0
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.25.9
virtualenv==20.0.10
wcwidth==0.1.8
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.11.2
